feat: text in ui camera
